Fix bug #24 and upgrade to ActiveMQ 5.14.3

+++ test/test_functionnal.py
@@ -43,7 +43,7 @@ class ActiveMQDockerTestCase(unittest.TestCase):
         global global_message
 
         # Run ActiveMQ container
-        rt,output = run_cmd('/usr/bin/docker ps | grep activemq')
+        rt,output = run_cmd(' /usr/bin/docker ps -f name=activemq')
         if rt == 0:
             print("Remove old ActiveMQ container")
             run_cmd('/usr/bin/docker stop activemq')
@@ -59,18 +59,21 @@ class ActiveMQDockerTestCase(unittest.TestCase):
 
 
         # Stop and remove ActiveMQ container
-        run_cmd('/usr/bin/docker stop activemq')
-        run_cmd('/usr/bin/docker rm activemq')
+        rt,output = run_cmd(' /usr/bin/docker ps -f name=activemq')
+        if rt == 0:
+            print("Remove old ActiveMQ container")
+            run_cmd('/usr/bin/docker stop activemq')
+            run_cmd('/usr/bin/docker rm activemq')
 
     def test_activemq_run(self):
 
         print("Start ActiveMQ")
-        rt,output = run_cmd('/usr/bin/docker run -d --name activemq webcenter/activemq:develop')
+        rt,output = run_cmd('/usr/bin/docker run -d --name activemq -p 61613:61613 webcenter/activemq:develop')
         if rt != 0:
             self.fail("Failed to start ActiveMQ container : %s" % output)
         time.sleep(60)
 
-        rt,output = run_cmd('/usr/bin/docker ps | grep activemq')
+        rt,output = run_cmd(' /usr/bin/docker ps -f name=activemq')
         self.assertEqual(rt, 0, "ActiveMQ container not running")
 
 
@@ -80,13 +83,13 @@ class ActiveMQDockerTestCase(unittest.TestCase):
         global global_error
 
         print("Start ActiveMQ")
-        rt,output = run_cmd('/usr/bin/docker run -d --name activemq webcenter/activemq:develop')
+        rt,output = run_cmd('/usr/bin/docker run -d --name activemq -p 61613:61613 webcenter/activemq:develop')
         if rt != 0:
             self.fail("Failed to start ActiveMQ container : %s" % output)
         time.sleep(60)
 
         # Init stomp connexion
-        conn = stomp.Connection([('activemq', 61613)])
+        conn = stomp.Connection([('172.17.0.1', 61613)])
         conn.set_listener('', MyListener())
         try:
           conn.start()
@@ -113,7 +116,7 @@ class ActiveMQDockerTestCase(unittest.TestCase):
 
         print("Start ActiveMQ")
         docker = '''
-/usr/bin/docker run --name='activemq' --rm -i -t -v /var/log/activemq:/var/log/activemq -p 8161:8161 \
+/usr/bin/docker run --name='activemq' -d -p 61613:61613 \
 -e 'ACTIVEMQ_REMOVE_DEFAULT_ACCOUNT=true' \
 -e 'ACTIVEMQ_ADMIN_LOGIN=admin' -e 'ACTIVEMQ_ADMIN_PASSWORD=admin_password' \
 -e 'ACTIVEMQ_USER_LOGIN=test' -e 'ACTIVEMQ_USER_PASSWORD=test_password' \
@@ -132,15 +135,15 @@ class ActiveMQDockerTestCase(unittest.TestCase):
 -e 'ACTIVEMQ_MIN_MEMORY=64' -e  'ACTIVEMQ_MAX_MEMORY=128' \
 -e 'ACTIVEMQ_ENABLED_SCHEDULER=true' \
 -e 'ACTIVEMQ_ENABLED_AUTH=true' \
-webcenter/activemq:test
+webcenter/activemq:develop
         '''
-        rt,output = run_cmd('docker run -d --name activemq webcenter/activemq:develop')
+        rt,output = run_cmd(docker)
         if rt != 0:
             self.fail("Failed to start ActiveMQ container : %s" % output)
         time.sleep(60)
 
         # Init stomp connexion
-        conn = stomp.Connection([('activemq', 61613)])
+        conn = stomp.Connection([('172.17.0.1', 61613)])
         conn.set_listener('', MyListener())
 
         try:
